+++ lib/crstemplate/crstemplate.go
@@ -7,10 +7,18 @@ package crstemplate
  * Curlrevshell template things
  * By J. Stuart McMurray
  * Created 20241205
- * Last Modified 20241205
+ * Last Modified 20241222
  */
 
-import _ "embed"
+import (
+	"bytes"
+	_ "embed"
+	"errors"
+	"fmt"
+	"io/fs"
+	"os"
+	"text/template"
+)
 
 // Default URL paths for comms with shells.
 const (
@@ -20,29 +28,94 @@ const (
 	DefaultURLPathScript = "c"  /* Script generation. */
 )
 
+// Subtemplate names in DefaultTemplate.  These may be overridden with the
+// {{define ...}} blocks in the template passed to -template.
+const (
+	SubtemplateCallback = "callback" /* "To get a shell" pastables */
+	SubtemplateFiles    = "files"    /* Static fileserver pastables */
+	SubtemplateScript   = "script"   /* Exceuted by /c */
+
+	// baseName isn't a subtemplate.  It's what's used as the name
+	// of the template outside {{define ...}} blocks.
+	baseName = "base"
+)
+
 // DefaultTemplate is the default callback script template.  It can be
 // overridden at runtime with -callback-template
 //
 //go:embed script.tmpl
 var DefaultTemplate string
 
-// Params are combined with the callback template to generate the callback
-// script.
-type Params struct {
-	PubkeyFP string /* Base64'd SHA256 hash of the server's TLS pubkey. */
-	URL      string /* URL to call back to the server. */
-	ID       string /* Random ID string. */
+// parsedDefaultTemplate is DefaultTemplate, parsed.
+var parsedDefaultTemplate = template.Must(newTemplate(DefaultTemplate))
 
-	// URLPaths are usually i, o, etc but may have been changed at
-	// compile-time.  They will not include slashes.
-	URLPaths URLPaths
+// ErrNoSubtemplates is returned from Execute when it is passed the name of a
+// template file which has no subtemplates defined.
+var ErrNoSubtemplates = errors.New("no subtemplates found")
+
+// Execute executes the given subtemplate with the given parameters.
+//
+// If file is not the empty string, it is taken as a template file to attempt
+// to read to override the built-in template.  Named subtemplates override
+// subtemplates with the similar name.
+//
+// If file is not the empty string and the file did not exist, DefaultTemplate
+// will be used but errors.Is(err, fs.ErrNotExist) will be true to enable a
+// warning that a template file is missing but the returned string will still
+// be the result of executing DefaultTemplate's subtemplate.
+func Execute(subtemplate, file string, params Params) (string, error) {
+
+	/* Try to parse the user's template if we have one. */
+	tmpl, uErr := mergeTemplateFrom(file)
+	if nil != uErr && !errors.Is(uErr, fs.ErrNotExist) {
+		return "", fmt.Errorf("adding templates: %w", uErr)
+	}
+
+	/* Execute the template. */
+	b := new(bytes.Buffer)
+	if err := tmpl.ExecuteTemplate(b, subtemplate, params); nil != err {
+		return "", fmt.Errorf("executing template: %w", err)
+	}
+
+	return b.String(), uErr /* Make be ENOENT. */
+
 }
 
-// URLPaths contain the parts of the URL paths indicating what an HTTPS
-// request is for.
-type URLPaths struct {
-	In     string /* Default: i */
-	InOut  string /* Default: io */
-	Out    string /* Default: o */
-	Script string /* Default: c */
+// mergeTemplateFrom clones parsedDefaultTemplate and Parses in the templates
+// from fn.  It returns an error if fn has no subtemplates defined.
+//
+// As a special case, mergeTemplateFrom returns parsedDefaultTemplate if fn is
+// the empty string or if fn could not be read (in which case the returned
+// error will also be non-nil).
+func mergeTemplateFrom(fn string) (*template.Template, error) {
+	/* If we're not actually reading a template, life's easy. */
+	if "" == fn {
+		return parsedDefaultTemplate, nil
+	}
+
+	/* Slurp the custom template. */
+	b, err := os.ReadFile(fn)
+	if nil != err {
+		return parsedDefaultTemplate, fmt.Errorf(
+			"reading template: %w",
+			err,
+		)
+	}
+
+	/* Make sure it parses as a template and has its own subtemplates. */
+	t, err := newTemplate(string(b))
+	if nil != err {
+		return nil, fmt.Errorf("parsing template: %w", err)
+	}
+	if 2 > len(t.Templates()) { /* Should always be at least one. */
+		return nil, ErrNoSubtemplates
+	}
+
+	/* Add to the default templates. */
+	return template.Must(parsedDefaultTemplate.Clone()).Parse(string(b))
+}
+
+// newTemplate returns a new template from s with the main template name baseName.
+func newTemplate(s string) (*template.Template, error) {
+	return template.New(baseName).Parse(s)
 }